# sipb-app/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SIPB UYM API",
    "version": "1.0.0",
    "description": "Dokumentasi API untuk Sistem Informasi Penemuan Barang (SIPB) UYM. Mencakup alur publik (pencarian, tracking, pelaporan mandiri via QR, AI Assistant) dan alur admin (dashboard, pengelolaan barang, aktivitas audit, manajemen user, ekspor laporan)."
  },
  "servers": [
    {
      "url": "/",
      "description": "Server lokal / default"
    }
  ],
...
  "paths": {
...
    "/admin/history": {
      "get": {
        "tags": ["Admin"],
        "summary": "Riwayat Barang Diambil (Admin)",
        "description": "Menampilkan daftar barang temuan yang sudah diserahkan ke pemiliknya (status: sudah_diambil).",
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Kata kunci pencarian riwayat (mencari nama barang, deskripsi, pelapor, NIM pelapor, pengambil, NIM pengambil)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Render halaman riwayat penyerahan dengan Inertia"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session",
        "description": "Autentikasi berbasis cookie session Laravel standar."
      }
    }
  }
}